Keep turn coordinator inside the architecture budget
The strict architecture gate regressed because safe-lane routing and event-sampling tests lived inline in turn_coordinator.rs. Move that test-only block into a dedicated child test file so the production module stays under the enforced size budget without changing runtime behavior or coverage intent.

Constraint: Governance enforces the 11200-line structural budget on crates/app/src/conversation/turn_coordinator.rs
Rejected: Extract runtime logic into a new production module | unnecessary behavior-bearing refactor for a budget-only regression
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: When turn_coordinator approaches its budget, move test-only coverage into child test modules before touching runtime code
Tested: cargo check -p loongclaw-app --all-features; LOONGCLAW_ARCH_STRICT=true ./scripts/check_architecture_boundaries.sh; git diff --cached --check
Not-tested: Full workspace clippy/test suite is still running separately for post-commit confirmation
